Move nested types to partial classes
