+++ src/DomainBlocks.Core/Subscriptions/EventStreamSubscription.ConsumerSessionGroup.cs
@@ -1,0 +1,79 @@
+using DomainBlocks.Core.Subscriptions.Concurrency;
+
+namespace DomainBlocks.Core.Subscriptions;
+
+public partial class EventStreamSubscription<TEvent, TPosition>
+{
+    private sealed class ConsumerSessionGroup : IDisposable
+    {
+        private readonly IReadOnlyDictionary<IEventStreamConsumer<TEvent, TPosition>, ConsumerSession> _sessions;
+
+        public ConsumerSessionGroup(
+            IEnumerable<IEventStreamConsumer<TEvent, TPosition>> consumers,
+            ArenaQueue<Notification> queue)
+        {
+            _sessions = consumers.ToDictionary(x => x, x => new ConsumerSession(x, queue));
+        }
+
+        public void Dispose()
+        {
+            foreach (var session in _sessions.Values)
+            {
+                session.Dispose();
+            }
+        }
+
+        public async Task<TPosition?> NotifyStarting(CancellationToken cancellationToken)
+        {
+            var tasks = _sessions.Values.Select(x => x.NotifyStarting(cancellationToken));
+            var startPositions = await Task.WhenAll(tasks);
+            return startPositions.Min();
+        }
+
+        public async Task NotifyCatchingUp(CancellationToken cancellationToken)
+        {
+            var tasks = _sessions.Values.Select(x => x.NotifyCatchingUp(cancellationToken));
+            await Task.WhenAll(tasks);
+
+            foreach (var session in _sessions.Values)
+            {
+                session.ResetCheckpointTimer();
+            }
+        }
+
+        public Task NotifyEvent(TEvent @event, TPosition position, CancellationToken cancellationToken)
+        {
+            var tasks = _sessions.Values.Select(x => x.NotifyEvent(@event, position, cancellationToken));
+            return Task.WhenAll(tasks);
+        }
+
+        public async Task NotifyLive(CancellationToken cancellationToken)
+        {
+            var notifyCheckpointTasks = _sessions.Values.Select(x => x.NotifyCheckpoint(cancellationToken));
+            await Task.WhenAll(notifyCheckpointTasks);
+
+            var notifyLiveTasks = _sessions.Values.Select(x => x.NotifyLive(cancellationToken));
+            await Task.WhenAll(notifyLiveTasks);
+
+            foreach (var session in _sessions.Values)
+            {
+                session.ResetCheckpointTimer();
+            }
+        }
+
+        public Task NotifySubscriptionDropped(
+            SubscriptionDroppedReason reason, Exception? exception, CancellationToken cancellationToken)
+        {
+            var tasks = _sessions.Values.Select(x => x.NotifySubscriptionDropped(reason, exception, cancellationToken));
+            return Task.WhenAll(tasks);
+        }
+
+        public async Task NotifyCheckpointTimerElapsed(
+            IEventStreamConsumer<TEvent, TPosition> consumer, CancellationToken cancellationToken)
+        {
+            var session = _sessions[consumer];
+            await session.NotifyCheckpoint(cancellationToken);
+            session.ResetCheckpointTimer();
+        }
+    }
+}
