+++ src/DomainBlocks.Core/Subscriptions/EventStreamSubscription.cs
@@ -2,7 +2,7 @@ using DomainBlocks.Core.Subscriptions.Concurrency;
 
 namespace DomainBlocks.Core.Subscriptions;
 
-public class EventStreamSubscription<TEvent, TPosition> :
+public partial class EventStreamSubscription<TEvent, TPosition> :
     IEventStreamSubscription,
     IEventStreamListener<TEvent, TPosition>
     where TPosition : struct, IEquatable<TPosition>, IComparable<TPosition>
@@ -11,7 +11,7 @@ public class EventStreamSubscription<TEvent, TPosition> :
 
     private readonly IEventStreamSubscriber<TEvent, TPosition> _subscriber;
     private readonly ConsumerSessionGroup _sessions;
-    private readonly ArenaQueue<Notification<TEvent, TPosition>> _queue = new(DefaultNotificationQueueSize);
+    private readonly ArenaQueue<Notification> _queue = new(DefaultNotificationQueueSize);
 
     private readonly TaskCompletionSource _eventLoopCompletedSignal =
         new(TaskCreationOptions.RunContinuationsAsynchronously);
@@ -100,7 +100,7 @@ public class EventStreamSubscription<TEvent, TPosition> :
         _eventLoopCompletedSignal.SetResult();
     }
 
-    private async Task HandleNotification(Notification<TEvent, TPosition> notification)
+    private async Task HandleNotification(Notification notification)
     {
         switch (notification.NotificationType)
         {
@@ -126,233 +126,6 @@ public class EventStreamSubscription<TEvent, TPosition> :
             default:
                 throw new ArgumentOutOfRangeException(
                     nameof(notification), $"Unknown notification type {notification.NotificationType}");
-        }
-    }
-
-    private sealed class ConsumerSession : IDisposable
-    {
-        private readonly IEventStreamConsumer<TEvent, TPosition> _consumer;
-        private readonly CheckpointTimer _checkpointTimer;
-        private TPosition? _startPosition;
-        private TPosition? _lastProcessedPosition;
-        private int _checkpointEventCount;
-        private CheckpointFrequency _checkpointFrequency = CheckpointFrequency.Default;
-
-        public ConsumerSession(
-            IEventStreamConsumer<TEvent, TPosition> consumer, ArenaQueue<Notification<TEvent, TPosition>> queue)
-        {
-            _consumer = consumer;
-            _checkpointTimer =
-                new CheckpointTimer(ct => queue.WriteAsync(x => x.SetCheckpointTimerElapsed(consumer), ct));
-        }
-
-        public void Dispose()
-        {
-            _checkpointTimer.Dispose();
-        }
-
-        public async Task<TPosition?> NotifyStarting(CancellationToken cancellationToken)
-        {
-            _startPosition = await _consumer.OnStarting(cancellationToken);
-            return _startPosition;
-        }
-
-        public Task NotifyCatchingUp(CancellationToken cancellationToken)
-        {
-            _checkpointFrequency = _consumer.CatchUpCheckpointFrequency;
-            return _consumer.OnCatchingUp(cancellationToken);
-        }
-
-        public async Task NotifyEvent(TEvent @event, TPosition position, CancellationToken cancellationToken)
-        {
-            if (_startPosition != null && position.CompareTo(_startPosition.Value) < 1)
-            {
-                // This consumer's start position is ahead of this event's position. Ignore.
-                return;
-            }
-
-            var isHandled = false;
-            var result = OnEventResult.Ignored;
-
-            do
-            {
-                try
-                {
-                    result = await _consumer.OnEvent(@event, position, cancellationToken);
-                    isHandled = true;
-                    break;
-                }
-                catch (Exception ex)
-                {
-                    var resolution = await _consumer.OnEventError(@event, position, ex, cancellationToken);
-
-                    switch (resolution)
-                    {
-                        case EventErrorResolution.Abort:
-                            throw;
-                        case EventErrorResolution.Retry:
-                            break;
-                        case EventErrorResolution.Skip:
-                            isHandled = true;
-                            break;
-                        default:
-                            throw new InvalidOperationException($"Unexpected EventErrorResolution value {resolution}");
-                    }
-                }
-            } while (!isHandled);
-
-            if (result == OnEventResult.Processed)
-            {
-                _lastProcessedPosition = position;
-                _checkpointEventCount += _checkpointFrequency.CanCheckpoint ? 1 : 0;
-
-                if (_checkpointEventCount == _checkpointFrequency.PerEventCount)
-                {
-                    await _consumer.OnCheckpoint(position, cancellationToken);
-                    _checkpointEventCount = 0;
-                    ResetCheckpointTimer();
-                }
-            }
-        }
-
-        public Task NotifyLive(CancellationToken cancellationToken)
-        {
-            _checkpointFrequency = _consumer.LiveCheckpointFrequency;
-            return _consumer.OnLive(cancellationToken);
-        }
-
-        public Task NotifySubscriptionDropped(
-            SubscriptionDroppedReason reason, Exception? exception, CancellationToken cancellationToken)
-        {
-            return _consumer.OnSubscriptionDropped(reason, exception, cancellationToken);
-        }
-
-        public async Task NotifyCheckpoint(CancellationToken cancellationToken)
-        {
-            if (_checkpointEventCount > 0)
-            {
-                await _consumer.OnCheckpoint(_lastProcessedPosition!.Value, cancellationToken);
-                _checkpointEventCount = 0;
-            }
-        }
-
-        public void ResetCheckpointTimer()
-        {
-            var duration = _checkpointFrequency.PerTimeInterval ?? Timeout.InfiniteTimeSpan;
-            _checkpointTimer.Reset(duration);
-        }
-    }
-
-    private sealed class ConsumerSessionGroup : IDisposable
-    {
-        private readonly IReadOnlyDictionary<IEventStreamConsumer<TEvent, TPosition>, ConsumerSession> _sessions;
-
-        public ConsumerSessionGroup(
-            IEnumerable<IEventStreamConsumer<TEvent, TPosition>> consumers,
-            ArenaQueue<Notification<TEvent, TPosition>> queue)
-        {
-            _sessions = consumers.ToDictionary(x => x, x => new ConsumerSession(x, queue));
-        }
-
-        public void Dispose()
-        {
-            foreach (var session in _sessions.Values)
-            {
-                session.Dispose();
-            }
-        }
-
-        public async Task<TPosition?> NotifyStarting(CancellationToken cancellationToken)
-        {
-            var tasks = _sessions.Values.Select(x => x.NotifyStarting(cancellationToken));
-            var startPositions = await Task.WhenAll(tasks);
-            return startPositions.Min();
-        }
-
-        public async Task NotifyCatchingUp(CancellationToken cancellationToken)
-        {
-            var tasks = _sessions.Values.Select(x => x.NotifyCatchingUp(cancellationToken));
-            await Task.WhenAll(tasks);
-
-            foreach (var session in _sessions.Values)
-            {
-                session.ResetCheckpointTimer();
-            }
-        }
-
-        public Task NotifyEvent(TEvent @event, TPosition position, CancellationToken cancellationToken)
-        {
-            var tasks = _sessions.Values.Select(x => x.NotifyEvent(@event, position, cancellationToken));
-            return Task.WhenAll(tasks);
-        }
-
-        public async Task NotifyLive(CancellationToken cancellationToken)
-        {
-            var notifyCheckpointTasks = _sessions.Values.Select(x => x.NotifyCheckpoint(cancellationToken));
-            await Task.WhenAll(notifyCheckpointTasks);
-
-            var notifyLiveTasks = _sessions.Values.Select(x => x.NotifyLive(cancellationToken));
-            await Task.WhenAll(notifyLiveTasks);
-
-            foreach (var session in _sessions.Values)
-            {
-                session.ResetCheckpointTimer();
-            }
-        }
-
-        public Task NotifySubscriptionDropped(
-            SubscriptionDroppedReason reason, Exception? exception, CancellationToken cancellationToken)
-        {
-            var tasks = _sessions.Values.Select(x => x.NotifySubscriptionDropped(reason, exception, cancellationToken));
-            return Task.WhenAll(tasks);
-        }
-
-        public async Task NotifyCheckpointTimerElapsed(
-            IEventStreamConsumer<TEvent, TPosition> consumer, CancellationToken cancellationToken)
-        {
-            var session = _sessions[consumer];
-            await session.NotifyCheckpoint(cancellationToken);
-            session.ResetCheckpointTimer();
-        }
-    }
-
-    private sealed class CheckpointTimer : IDisposable
-    {
-        private readonly Func<CancellationToken, Task> _onElapsed;
-        private TimeSpan _duration = Timeout.InfiniteTimeSpan;
-        private CancellationTokenSource _cancellationTokenSource = new();
-        private Task _task;
-
-        public CheckpointTimer(Func<CancellationToken, Task> onElapsed)
-        {
-            _onElapsed = onElapsed;
-            _task = Task.Delay(_duration, _cancellationTokenSource.Token);
-        }
-
-        public void Reset(TimeSpan duration)
-        {
-            _duration = duration;
-
-            if (!_task.IsCompletedSuccessfully || !_cancellationTokenSource.TryReset())
-            {
-                _cancellationTokenSource.Cancel();
-                _cancellationTokenSource.Dispose();
-                _cancellationTokenSource = new CancellationTokenSource();
-            }
-
-            _task = Task
-                .Delay(_duration, _cancellationTokenSource.Token)
-                .ContinueWith(async t =>
-                {
-                    await t;
-                    await _onElapsed(_cancellationTokenSource.Token);
-                });
-        }
-
-        public void Dispose()
-        {
-            _cancellationTokenSource.Cancel();
-            _cancellationTokenSource.Dispose();
         }
     }
