--- src/DomainBlocks.Core/Subscriptions/NotificationType.cs
@@ -1,10 +1,0 @@
-namespace DomainBlocks.Core.Subscriptions;
-
-internal enum NotificationType
-{
-    CatchingUp,
-    Event,
-    Live,
-    SubscriptionDropped,
-    CheckpointTimerElapsed
-}
